refactor(vue): organize package sources
